requirements.txt
# -----------------------------
# 🧠 Core Framework
# -----------------------------
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1

# -----------------------------
# ☁️ Database (MongoDB)
# -----------------------------
motor==3.6.0
pymongo==4.9.1
certifi==2024.8.30

# -----------------------------
# 🖼️ Image & ML Processing
# -----------------------------
Pillow==10.4.0
ultralytics==8.0.196
torch>=1.8.0
torchvision>=0.9.0

# -----------------------------
# 🤖 AI Integration (Gemini API)
# -----------------------------
google-generativeai==0.8.3

# -----------------------------
# ☁️ AWS S3 Integration
# -----------------------------
boto3==1.35.0

# -----------------------------
# 🌐 File Uploads
# -----------------------------
python-multipart==0.0.9

# -----------------------------
# 🧪 Development Utilities
# -----------------------------
httpx==0.27.2
requests==2.32.3
